// frontend/src/app/core/api/api/models/carDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CarDto { 
    carId?: number;
    licensePlate?: string;
    brand?: string;
    model?: string;
    constructionYear?: string;
    color?: string;
    chassisNumber?: string;
    mileage?: number;
    status?: string;
    category?: string;
    dailyPrice?: number;
    image64?: string;
}